资料/底盘/qyh_standard_api/requirements.txt
# Standard Robots Matrix API Client
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.21.0
PyYAML>=6.0
